Refactor to make map a seperate react component
