+++ frontend/src/components/Home.js
@@ -1,0 +1,10 @@
+import React from "react";
+
+const Home = () => (
+  <div className="content">
+    <h2>Welcome to the Mapping Application</h2>
+    <p>Explore locations with detailed metadata on the interactive map.</p>
+  </div>
+);
+
+export default Home;
